Full vargs printk.

+++ src/include/stdint.h
@@ -3,13 +3,13 @@
 
 typedef char 			int8_t;
 typedef short int 		int16_t;
-typedef long int		int32_t;
-typedef long long int		int64_t;
+typedef int			int32_t;
+typedef long int		int64_t;
 
 typedef unsigned char 		uint8_t;
 typedef unsigned short int 	uint16_t;
-typedef unsigned long int 	uint32_t;
-typedef unsigned long long int 	uint64_t;
+typedef unsigned int	 	uint32_t;
+typedef unsigned long int 	uint64_t;
 
 typedef uint64_t 		size_t;
 typedef int64_t 		ssize_t;
